Three calls that asked nothing stop asking
glValidateProgram ran right after linking and nobody read GL_VALIDATE_STATUS
— and reading it would have said nothing either: the call grades a program
against the current GL state, and directly after a link there are no bound
textures, samplers or vertex array to grade it against.

GetShader(size_t) and GetNumberShaders existed for one caller, which used
them to walk every shader and set the camera matrices. Shaders now offers
SetCameraUniforms and keeps the walk where the shaders live; two methods
leave the interface with it.

GetNumberAttributes asked the driver for a count the vector beside it
already held. The GL query stays where it is needed — in the gather, before
the vector is filled — and the accessor answers from the vector.

+++ src/infrastructure/graphics/shaders_info.hpp
@@ -62,9 +62,9 @@ class ShaderInfos {
 
   [[nodiscard]] const std::vector<ShaderInfo>& GetAttributesInfos() const;
 
-  [[nodiscard]] int GetNumberAttributes() const;
+  [[nodiscard]] size_t GetNumberAttributes() const;
 
-  [[nodiscard]] int GetNumberUniforms() const;
+  [[nodiscard]] size_t GetNumberUniforms() const;
 
   void PrintAttributesInfo() const;
